FIX: Update image

+++ triply-backend/app/Http/Controllers/Admin/TripController.php
@@ -89,7 +89,9 @@ class TripController extends Controller
         $trip->category_id = $data['category_id'];
         $trip->description = $data['description'];
         if (array_key_exists('image', $data)) {
-            Storage::delete($trip->cover_image);
+            if ($trip->cover_image) {
+                Storage::delete($trip->cover_image);
+            }
             $img_url = Storage::putFile('trips', $data['image']);
             $trip->cover_image = $img_url;
         }
